vdk-properties-fs: new plugin for local FS properties storage
We need a simplistic plugin, that allows a developer or presenter
to quickly store secrets on the local FS during a dev/demo session,
that does not require a prerequisite of the entire Control Service
installed and running.

Added a FS properties client, that uses a customizable dir and filename
on the local file system to store the properties key-values. It does
support multiple teams and data jobs, the properties of those are nested
under a dedicated key for deduplication.
JSON format is used, enabling nested data structures.

Testing Done: added `test_fs_properties_plugin.py`, that covers multiple
data jobs using matching keys, and default/customized directory and
filename for file storage on the local FS.

+++ projects/vdk-plugins/vdk-properties-fs/src/vdk/plugin/properties_fs/fs_properties_client.py
@@ -1,9 +1,10 @@
 # Copyright 2021 VMware, Inc.
 # SPDX-License-Identifier: Apache-2.0
+import json
 import logging
 import os
-from configparser import ConfigParser
 from typing import Dict
+from typing import Optional
 
 from vdk.api.plugin.plugin_input import IPropertiesServiceClient
 
@@ -13,46 +14,51 @@ log = logging.getLogger(__name__)
 class FileSystemPropertiesServiceClient(IPropertiesServiceClient):
     """
     Implementation of IProperties that are kept on local FS only. Stateless.
-    A single properties file store, supporting multiple teams and data jobs properties,
-    via placing the key-values in a dedicated team and job name section to deduplicate.
+    A single JSON file store, supporting multiple teams and data jobs properties,
+    via nesting the key-values under a dedicated team and job name key to deduplicate.
     """
 
     def __init__(self, directory: str, filename: str):
         self._file_path = os.path.join(directory, filename)
 
     def read_properties(self, job_name: str, team_name: str) -> Dict:
-        config_parser = ConfigParser()
-        config_parser.read(self._file_path)
+        if os.path.isfile(self._file_path) and os.stat(self._file_path).st_size != 0:
+            props: Dict
+            with open(self._file_path) as props_file:
+                props = json.load(props_file)
 
-        # filter properties by prefix for this particular team and data job
-        prefix = FileSystemPropertiesServiceClient._prefix(team_name, job_name)
-        if config_parser.has_section(prefix):
-            return dict(config_parser.items(prefix))
+            # filter properties by prefix for this particular team and data job
+            prefix = FileSystemPropertiesServiceClient._prefix(team_name, job_name)
+            if props.keys().__contains__(prefix):
+                return props.get(prefix)
         return {}
 
     def write_properties(self, job_name: str, team_name: str, properties: Dict) -> Dict:
         log.warning(
             "You are using local FS Properties client. "
-            "That means the properties will not be stored using the Control Service."
+            "That means the properties will not be stored using the Control Service, yet FS storage is stateful and "
+            "all properties data will survive VDK process restarts."
             "The properties will be read and written to the local file system only."
         )
-        config_parser = ConfigParser()
-        config_parser.read(self._file_path)
+        props: Optional[Dict] = None
 
-        # filter properties by prefix for this particular team and data job
-        prefix = FileSystemPropertiesServiceClient._prefix(team_name, job_name)
-        # overwrite section to update its properties
-        if config_parser.has_section(prefix):
-            config_parser.remove_section(prefix)
-        config_parser.add_section(prefix)
+        # init storage file
+        if not os.path.isfile(self._file_path) or os.stat(self._file_path).st_size == 0:
+            with open(self._file_path, "w+") as props_file:
+                props = {}
+                json.dump(props, props_file)
 
-        # add section properties
-        for k, v in properties.items():
-            config_parser.set(prefix, k, v)
+        with open(self._file_path, "r+") as props_file:
+            props = json.load(props_file) if props is None else props
 
-        # overwrite file storage with updated values
-        with open(self._file_path, "w+") as props_file:
-            config_parser.write(props_file)
+            # filter properties by prefix for this particular team and data job
+            prefix = FileSystemPropertiesServiceClient._prefix(team_name, job_name)
+            # overwrite properties
+            props_file.seek(0)
+            props.update({prefix: properties})
+            props_file.truncate()
+            # store
+            json.dump(props, props_file)
         return properties
 
     @staticmethod
